Add tests for Stylelint
