+++ .eslintrc.json
@@ -1,6 +1,7 @@
 {
   "env": {
     "es6": true,
+    "jest": true,
     "mocha": true,
     "node": true
   },
