+++ .prettierrc
@@ -1,4 +1,5 @@
 {
-  "singleQuote": true,
-  "printWidth": 100
+  "endOfLine": "auto",
+  "printWidth": 100,
+  "singleQuote": true
 }
