+++ docs/rules/css/no-custom-font.md
@@ -1,0 +1,5 @@
+# Use a standard font
+
+## Rule Details
+
+TODO
