+++ docs/rules/js/externalize-js-css.md
