+++ docs/rules/js/no-try-catch-finally.md
