+++ package.json
@@ -9,10 +9,11 @@
   ],
   "main": "lib/index.js",
   "scripts": {
-    "test:js": "mocha tests --full-trace --recursive ",
-    "test:css": "npx stylelint \"resources/css/*.css\"",
-    "report:js": "eslint resources/js/* -f json > eslint-report.json",
-    "report:css": "stylelint resources/css/* -f json > stylelint-report.json"
+    "test": "npm run test:js && npm run test:css",
+    "test:js": "mocha tests/lib/rules/js --full-trace --recursive",
+    "test:css": "jest tests/lib/rules/css",
+    "report:js": "eslint resources/js/**/*.js -f json > eslint-report.json",
+    "report:css": "stylelint resources/css/**/*.css -f json > stylelint-report.json"
   },
   "files": [
     "lib"
@@ -31,12 +32,15 @@
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-self": "^1.2.1",
+    "jest": "^28.1.0",
+    "jest-preset-stylelint": "^5.0.3",
     "mocha": "^10.0.0",
     "prettier": "^2.2.1",
     "stylelint": "^14.8.5",
+    "stylelint-config-prettier": "^9.0.3",
     "stylelint-config-recommended": "^7.0.0",
     "stylelint-config-standard": "^25.0.0",
-    "stylelint-group-selectors": "^1.0.8"
+    "stylelint-prettier": "^2.0.0"
   },
   "engines": {
     "node": ">=0.10.0"
