--- resources/css/EfficiencyCssSelector.css
@@ -1,16 +1,0 @@
-.foo .bar .baz .lorem {}
-
-.foo .baz {
-    &>.bar .lorem {}
-}
-
-div {}
-
-.foo div {}
-
-#foo #bar>#baz {}
-
-.foo+div :not (a b ~ c) {}
-
-
-/* `a b ~ c` is inside `:not()`, so it is evaluated separately */
