--- resources/css/GroupCssDeclaration.css
@@ -1,15 +1,0 @@
-.a {
-    display: inline-block;
-    width: 100px;
-}
-
-.b {
-    display: inline-block;
-    width: 100px;
-}
-
-.a,
-.b {
-    display: inline-block;
-    width: 100px;
-}
