--- resources/css/PromoteStandardPolicy.css
@@ -1,23 +1,0 @@
-@font-face {
-    font-family: 'quadranta';
-    src: url('quadranta.eot?') format('eot'), url('quadranta.otf') format('truetype'), url('quadranta.woff') format('woff'), url('quadranta.svg#QuadrantaBold') format('svg');
-    font-weight: normal;
-    font-style: normal;
-}
-
-h4 {
-    font-family: quadranta, sans-serif;
-    font-size: 3em;
-}
-
-.exempleserif {
-    font-family: Times, "Times New Roman", Georgia, serif;
-}
-
-.exemplesansserif {
-    font-family: sans-serif;
-}
-
-.blueColor {
-    color: blue;
-}
