+++ resources/css/no-custom-font.css
@@ -1,0 +1,12 @@
+@font-face {
+  font-family: quadranta;
+  src: url('quadranta.eot?') format('eot'), url('quadranta.otf') format('truetype'),
+    url('quadranta.woff') format('woff'), url('quadranta.svg#QuadrantaBold') format('svg');
+  font-weight: normal;
+  font-style: normal;
+}
+
+h4 {
+  font-family: quadranta, sans-serif;
+  font-size: 3em;
+}
