+++ tests/lib/rules/css/rule-tester.js
@@ -1,0 +1,6 @@
+const { getTestRule } = require('jest-preset-stylelint');
+
+module.exports.createTestRule = (pluginName) => ({
+  ruleFilePath: `../../../../lib/rules/css/${pluginName}.js`,
+  testRule: getTestRule({ plugins: [`./lib/rules/css/${pluginName}.js`] }),
+});
